+++ plugins/org.ietr.preesm.mapper/src/org/ietr/preesm/mapper/scenariogen/ScenarioGenerator.java
@@ -152,7 +152,7 @@ public class ScenarioGenerator extends AbstractTaskImplementation {
 						((PreesmScenario) outputs.get("scenario"))
 								.getTimingManager().setTiming(sdfV.getName(),
 										op.getComponent().getVlnv().getName(),
-										Integer.parseInt(timeStr));
+										Long.parseLong(timeStr));
 					}
 				}
